archiving emulationStation "service" for "Goo-e games\List"
since the art scrapper shipped with EmulationStation is antiquated and extremely slow, I'm abandoning support for it in the games launcher. Maybe I'll come back to it someday, so this branch will serve as an archive copy of the work that was done so far (believe it is currently broken anyway).